+++ src/lib/transformer/md-callout-task.ts
@@ -1,5 +1,5 @@
 import { MD_Template } from "../md-template";
-import { MD_Exporter_Parameter_Type } from "../md-exporter";
+import { MD_Transporter_Parameter_Type } from "../md-transporter";
 import { MD_FileContent_Interface } from "../md-filesystem";
 import { MD_Observer_Interface } from "../md-observer";
 import {
@@ -109,7 +109,7 @@ export class MD_Callout_Transformer extends MD_Transformer_AbstractBase {
     this.parameter = parameter;
   }
 
-  public set_job_parameter(job_paramter: MD_Exporter_Parameter_Type): void {
+  public set_job_parameter(job_paramter: MD_Transporter_Parameter_Type): void {
     super.set_job_parameter(job_paramter);
   }