+++ src/lib/transformer/md-remove-todos-task.ts
@@ -1,5 +1,5 @@
 import { MD_FileContent_Interface } from "../md-filesystem";
-import { MD_Exporter_Parameter_Type } from "../md-exporter";
+import { MD_Transporter_Parameter_Type } from "../md-transporter";
 import { MD_Observer_Interface } from "../md-observer";
 import {
   MD_Transformer_AbstractBase,
@@ -31,7 +31,7 @@ export class MD_RemoveTODOS_Transformer extends MD_Transformer_AbstractBase {
     this.observer_subject.add_observer(observer);
   }
 
-  public set_job_parameter(job_paramter: MD_Exporter_Parameter_Type): void {
+  public set_job_parameter(job_paramter: MD_Transporter_Parameter_Type): void {
     super.set_job_parameter(job_paramter);
     // Das ist ein Hack.
     //? Eigentlich ist die Metode ja in der abstrakten Basisklasse vorhanden.