+++ test/md-export-1.test.ts
@@ -1,8 +1,8 @@
-import { MD_Exporter } from "../src/lib/md-exporter";
+import { MD_Transporter } from "../src/lib/md-transporter";
 
 /**
  * This test deals with the sharing of a longform document 
  * via a config file.
  */
-const my_exporter: MD_Exporter = new MD_Exporter();
-my_exporter.perform_job_from("test-data-obsidian-vault/transport-config.json", "Example-Job No.1");
+const my_transporter: MD_Transporter = new MD_Transporter();
+my_transporter.perform_job_from("test-data-obsidian-vault/transport-config.json", "Example-Job No.1");